refactor: reogranization and renaming
